+++ src/Altinn.Broker.API/Models/FileTransferInitializeExt.cs
@@ -73,65 +73,3 @@ public class FileTransferInitalizeExt
     [JsonPropertyName("disableVirusScan")]
     public bool? DisableVirusScan { get; set; } = false;
 }
-internal class MD5ChecksumAttribute : ValidationAttribute
-{
-    public MD5ChecksumAttribute()
-    {
-    }
-
-    protected override ValidationResult IsValid(object? value, ValidationContext validationContext)
-    {
-        var stringValue = value as string;
-        if (string.IsNullOrWhiteSpace(stringValue))
-        {
-            return ValidationResult.Success!;
-        }
-        if (stringValue.Length != 32)
-        {
-            return new ValidationResult("The checksum, if used, must be a MD5 hash with a length of 32 characters");
-        }
-        if (stringValue.ToLowerInvariant() != stringValue)
-        {
-            return new ValidationResult("The checksum, if used, must be a MD5 hash in lower case");
-        }
-        return ValidationResult.Success!;
-    }
-}
-
-[AttributeUsage(AttributeTargets.Property)]
-internal class PropertyListAttribute : ValidationAttribute
-{
-    public PropertyListAttribute()
-    {
-    }
-
-    protected override ValidationResult IsValid(object value, ValidationContext validationContext)
-    {
-        if (value == null)
-        {
-            return ValidationResult.Success;
-        }
-
-        if (!(value is Dictionary<string, string>))
-        {
-            return new ValidationResult("PropertyList Object is not of proper type");
-        }
-
-        var dictionary = (Dictionary<string, string>)value;
-
-        if (dictionary.Count > 10)
-            return new ValidationResult("PropertyList can contain at most 10 properties");
-
-        foreach (var keyValuePair in dictionary)
-        {
-            if (keyValuePair.Key.Length > 50)
-                return new ValidationResult(String.Format("PropertyList Key can not be longer than 50. Length:{0}, KeyValue:{1}", keyValuePair.Key.Length.ToString(), keyValuePair.Key));
-
-            if (keyValuePair.Value.Length > 300)
-                return new ValidationResult(String.Format("PropertyList Value can not be longer than 300. Length:{0}, Value:{1}", keyValuePair.Value.Length.ToString(), keyValuePair.Value));
-        }
-
-        return ValidationResult.Success;
-    }
-}
-
