+++ src/Altinn.Broker.API/ValidationAttributes/PropertyListAttribute.cs
@@ -1,0 +1,41 @@
+using System.ComponentModel.DataAnnotations;
+
+namespace Altinn.Broker.API.Models
+{
+    [AttributeUsage(AttributeTargets.Property)]
+    internal class PropertyListAttribute : ValidationAttribute
+    {
+        public PropertyListAttribute()
+        {
+        }
+
+        protected override ValidationResult IsValid(object? value, ValidationContext validationContext)
+        {
+            if (value == null)
+            {
+                return ValidationResult.Success!;
+            }
+
+            if (!(value is Dictionary<string, string>))
+            {
+                return new ValidationResult("PropertyList Object is not of proper type");
+            }
+
+            var dictionary = (Dictionary<string, string>)value;
+
+            if (dictionary.Count > 10)
+                return new ValidationResult("PropertyList can contain at most 10 properties");
+
+            foreach (var keyValuePair in dictionary)
+            {
+                if (keyValuePair.Key.Length > 50)
+                    return new ValidationResult(String.Format("PropertyList Key can not be longer than 50. Length:{0}, KeyValue:{1}", keyValuePair.Key.Length.ToString(), keyValuePair.Key));
+
+                if (keyValuePair.Value.Length > 300)
+                    return new ValidationResult(String.Format("PropertyList Value can not be longer than 300. Length:{0}, Value:{1}", keyValuePair.Value.Length.ToString(), keyValuePair.Value));
+            }
+
+            return ValidationResult.Success!;
+        }
+    }
+}
